refactor: reorganize and rename context providers
