+++ src/assets/custom/icon-dir.tsx
@@ -1,6 +1,6 @@
 import { type SVGProps } from 'react'
 import { cn } from '@/lib/utils'
-import { type Direction } from '@/context/direction-context'
+import { type Direction } from '@/context/direction-provider'
 
 type IconDirProps = SVGProps<SVGSVGElement> & {
   dir: Direction
