+++ src/features/tasks/components/data-table-row-actions.tsx
@@ -15,9 +15,9 @@ import {
   DropdownMenuSubTrigger,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu'
-import { useTasks } from '../context/tasks-context'
 import { labels } from '../data/data'
 import { taskSchema } from '../data/schema'
+import { useTasks } from './tasks-provider'
 
 type DataTableRowActionsProps<TData> = {
   row: Row<TData>
