+++ src/features/tasks/index.tsx
@@ -4,11 +4,11 @@ import { Main } from '@/components/layout/main'
 import { ProfileDropdown } from '@/components/profile-dropdown'
 import { Search } from '@/components/search'
 import { ThemeSwitch } from '@/components/theme-switch'
-import { columns } from './components/columns'
-import { DataTable } from './components/data-table'
+import { TasksTable } from './components/data-table'
+import { tasksColumns } from './components/tasks-columns'
 import { TasksDialogs } from './components/tasks-dialogs'
 import { TasksPrimaryButtons } from './components/tasks-primary-buttons'
-import { TasksProvider } from './context/tasks-context'
+import { TasksProvider } from './components/tasks-provider'
 import { tasks } from './data/tasks'
 
 export function Tasks() {
@@ -34,7 +34,7 @@ export function Tasks() {
           <TasksPrimaryButtons />
         </div>
         <div className='-mx-4 flex-1 overflow-auto px-4 py-1 lg:flex-row lg:space-y-0 lg:space-x-12'>
-          <DataTable data={tasks} columns={columns} />
+          <TasksTable data={tasks} columns={tasksColumns} />
         </div>
       </Main>
