+++ src/features/users/components/users-primary-buttons.tsx
@@ -1,6 +1,6 @@
 import { MailPlus, UserPlus } from 'lucide-react'
 import { Button } from '@/components/ui/button'
-import { useUsers } from '../context/users-context'
+import { useUsers } from './users-provider'
 
 export function UsersPrimaryButtons() {
   const { setOpen } = useUsers()
